OTel.Instrumentation.Persistent.Internal

Disclaimer

In general, changes to this module will not be reflected in the library's version updates. Direct use of this module should be done with utmost care, otherwise invariants will easily be violated.

buildSqlPoolHooks :: forall m. (MonadIO m, MonadMask m) => TracerProvider -> IO (SqlPoolHooks m SqlBackend)

sqlPoolHooks :: forall m. (MonadIO m, MonadMask m) => TracingBackend -> SqlPoolHooks m SqlBackend

setNewAlterBackend :: TracingBackend -> SqlPoolHooks m SqlBackend -> SqlPoolHooks m SqlBackend

setNewRunBefore :: forall m. (MonadIO m, MonadMask m) => TracingBackend -> SqlPoolHooks m SqlBackend -> SqlPoolHooks m SqlBackend

setNewRunAfter :: forall m. (MonadIO m, MonadMask m) => TracingBackend -> SqlPoolHooks m SqlBackend -> SqlPoolHooks m SqlBackend

setNewRunOnException :: forall m. (MonadIO m, MonadMask m) => TracingBackend -> SqlPoolHooks m SqlBackend -> SqlPoolHooks m SqlBackend

modifyRunOnException :: forall m backend. SqlPoolHooks m backend -> ((backend -> Maybe IsolationLevel -> SomeException -> m ()) -> backend -> Maybe IsolationLevel -> SomeException -> m ()) -> SqlPoolHooks m backend

modifyGetStatement :: SqlBackendHooks -> ((SqlBackend -> Text -> Statement -> IO Statement) -> SqlBackend -> Text -> Statement -> IO Statement) -> SqlBackendHooks

unsafeTracedAcquire :: TracingBackend -> SpanSpec -> Acquire a -> Acquire a

persistValuesToAttrs :: [PersistValue] -> AttrsBuilder 'AttrsForSpan

newtype Bytes

Constructors

Bytes

Fields

unBytes :: ByteString